Clean up test logic of OAuth2 and OpenID providers
